+++ src/editor/editor.cpp
@@ -2,14 +2,18 @@
 #include <Geode/modify/EditorUI.hpp>
 #include <Geode/modify/LevelEditorLayer.hpp>
 #include <utils/include.hpp>
+#include <alphalaneous.tinker/include/UIScaling.hpp>
 
 using namespace geode::prelude;
 
 namespace nwo5::editor {
     namespace impl {
-        static bool s_shouldMoveObject;
+        static bool s_shouldMoveObject = true;
 
-        static bool s_editButtonsLoaded;
+        static std::optional<float> s_tinkerUIScale = std::nullopt;
+        static std::optional<float> s_beUIScale = std::nullopt;
+
+        static bool s_editButtonsLoaded = false;
         static std::vector<EditTabButton> s_editTabButtons;
         static std::unordered_map<std::string, CCMenuItemSpriteExtra*> s_editTabButtonMap;
         static std::unordered_set<std::string> s_removedEditTabButtons;
@@ -166,6 +170,15 @@ namespace nwo5::editor {
 
             bool init(LevelEditorLayer* editorLayer) {
                 s_editorUIEarlyPtr = this;
+
+                if (auto mod = Loader::get()->getLoadedMod("alphalaneous.tinker")) {
+                    s_tinkerUIScale = mod->getSettingValue<bool>("UIScaling-enabled")
+                        ? mod->getSettingValue<float>("UIScaling-scale")
+                        : 1.0f;
+                }
+                if (auto mod = Loader::get()->getLoadedMod("hjfod.betteredit")) {
+                    s_beUIScale = mod->getSettingValue<float>("scale-factor");
+                }
 
                 return EditorUI::init(editorLayer);
             }
@@ -478,6 +491,14 @@ namespace nwo5::editor {
         }
         
         impl::getFakePauseLayer()->saveLevel();
+    }
+
+    float uiScale() {
+        static auto _ = tinker::api::ui_scaling::UIScaleUpdated().listen([] (float pScale, bool, bool) {
+            impl::s_tinkerUIScale = pScale;
+        });
+
+        return impl::s_beUIScale.value_or(impl::s_tinkerUIScale.value_or(1.0f));
     }
 
     bool registerEditTabButton(impl::EditTabButton::SpriteFunc pSprite, std::string pKey, float pScale, int pPrio, impl::EditTabButton::Callback pCallback) {